changed range unit in the project to realtime from sample. refined undo display

+++ src/data/region.rs
@@ -50,7 +50,7 @@ pub enum Content {
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct Region {
     /// range stores a real time, not in sample.
-    pub range: AtomicRange<i64>,
+    pub range: AtomicRange<f64>,
     pub content: Content,
     pub label: String,
 }
@@ -58,7 +58,7 @@ pub struct Region {
 impl Region {
     /// Utility function that converts a raw region into the region with fadein/out transformer.
     ///
-    pub fn new(range: AtomicRange<i64>, content: Content, label: impl Into<String>) -> Self {
+    pub fn new(range: AtomicRange<f64>, content: Content, label: impl Into<String>) -> Self {
         Self {
             range,
             content,
@@ -67,7 +67,7 @@ impl Region {
     }
     pub fn with_fade(origin: Self) -> Self {
         Self::new(
-            AtomicRange::<i64>::new(origin.range.start(), origin.range.end()),
+            AtomicRange::<f64>::new(origin.range.start(), origin.range.end()),
             Content::Transformer(
                 RegionFilter::FadeInOut(Arc::new(FadeParam {
                     time_in: 0.1.into(),
@@ -123,7 +123,7 @@ impl Region {
 impl std::default::Default for Region {
     fn default() -> Self {
         Self {
-            range: AtomicRange::<i64>::new(0, 0),
+            range: AtomicRange::<f64>::new(0.0, 0.0),
             content: Content::Generator(Generator::default()),
             label: "".to_string(),
         }
